Make era-test-node a library (#79)
* Make era-test-node a library

* Fixed merge changes

* fixed tests

* fmt

* Moved showcalls to node
